ref stock handlers
